requirements.txt
crochet==1.12.0
Flask==1.1.2
Scrapy==2.1.0
SQLAlchemy==1.3.17
bibtexparser~=1.1.0
alembic~=1.4.2
Flask-Classful==0.14.2
Sphinx==3.1.2
python-doi==0.2.0